backend/api/views.py
from django.shortcuts import render
from rest_framework.decorators import api_view
from rest_framework.response import Response

@api_view(['POST'])
def autocorrect_text(request):
    text = request.data.get('text', '')
    # Implement your RAG logic here
    corrected_text = text  # Replace with actual autocorrect logic
    return Response({'corrected_text': corrected_text})